Update analysis.correlate
Updated analysis.correlate to take ref_ws_col and site_ws_col arguments. This should allow the package to be more flexible and removes the need to have each column labeled ref and site. This came up when trying to use anemoi to analyze EIA data and wanting to correlate between NCF and MERRA.